third_party/dawn/third_party/catapult/third_party/vinn/third_party/parse5/test/data/tree_construction_options/decode_html_entities.dat

#data
<html><head></head><body>&nbsp;</body></html>
#disable-html-entities-decoding
true
#errors
#document
| <html>
|   <head>
|   <body>
|     "&nbsp;"

#data
<html><head></head><body><a href="&amp;test&copy;"></a></body></html>
#disable-html-entities-decoding
true
#errors
#document
| <html>
|   <head>
|   <body>
|     <a>
|       href="&amp;test&copy;"
